Add projects page, common components refactor
