+++ src/components/Header.astro
@@ -24,6 +24,8 @@ import WebsiteIcon from "../assets/icons/website.png";
     <nav class="flex items-center gap-1" aria-label="Primary navigation">
       <Link href="/blog" underline>blog</Link>
       <span aria-hidden="true">/</span>
+      <Link href="/projects" underline>projects</Link>
+      <span aria-hidden="true">/</span>
       <Link href="/about" underline>about</Link>
     </nav>
   </div>
