+++ src/components/IconLink.astro
@@ -11,7 +11,7 @@ const props = Astro.props;
   target="_blank"
   rel="noopener noreferrer"
   underline
-  class="inline-flex items-center gap-2 rounded-sm border border-neutral-200 px-4 py-2 dark:border-neutral-700"
+  class="inline-flex items-center gap-2 rounded-md border border-neutral-200 px-4 py-2 dark:border-neutral-700"
   {...props}
 >
   <slot />
