+++ src/components/NewsletterSubscribe.astro
@@ -1,0 +1,18 @@
+---
+import IconLink from "./IconLink.astro";
+import Mail from "../assets/icons/mail.svg";
+---
+
+<div class="not-prose space-y-1.5">
+  <h2 class="font-semibold">Want to receive updates straight in your inbox?</h2>
+
+  <IconLink href="https://newsletter.chornonoh-vova.com/subscribe">
+    <Mail
+      width={18}
+      height={18}
+      class="stroke-emerald-600"
+      role="presentation"
+    />
+    Subscribe to the newsletter
+  </IconLink>
+</div>
