+++ src/components/PostCard.astro
@@ -11,7 +11,7 @@ const { entry } = Astro.props;
 
 <a
   href={`/${entry.collection}/${entry.id}`}
-  class="group flex flex-nowrap items-center gap-2 rounded-lg border border-neutral-200 py-3 pr-3 pl-5 hover:text-black dark:border-neutral-700 dark:hover:text-white"
+  class="group flex flex-nowrap items-center gap-2 rounded-xl border border-neutral-200 py-3 pr-3 pl-5 hover:text-black dark:border-neutral-700 dark:hover:text-white"
 >
   <div class="flex flex-1 flex-col gap-1 overflow-x-hidden">
     <h3 class="font-semibold">
